Refactor IOUtils to proper package
